requirements.txt
Flask==3.1.0
pymongo==4.12.0
python-dotenv==1.1.0
Werkzeug==3.1.3
temporalio==1.15.0